Renamed utils.h to utils.hpp
